app/requirements.txt
flask
flask_mongoengine
dependencies/dist/Flask_User-1.0.2.3-py2.py3-none-any.whl
flask_wtf
wtforms
flask-restx
email_validator
flask-babel==2.0.0
python-dotenv==0.15.0
requests==2.25.1
tqdm==4.59.0
numpy==1.20.1
pandas==1.2.3
